+++ arts.php
@@ -64,7 +64,7 @@ switch ($_REQUEST['action']) {
         $image_data = Art::get_from_source($data, $object_type);
 
         // If we got something back insert it
-        if ($image_data !== null) {
+        if ($image_data !== '') {
             $art = new Art($object_id, $object_type);
             $art->insert($image_data, $_FILES['file']['type']);
             show_confirmation(T_('No Problem'), T_('Art has been added'), $burl);
@@ -91,7 +91,7 @@ switch ($_REQUEST['action']) {
             $upload['mime'] = 'image/' . $path_info['extension'];
             $image_data     = Art::get_from_source($upload, $object_type);
 
-            if ($image_data !== null) {
+            if ($image_data['object_id']) {
                 $art->insert($image_data, $upload['0']['mime']);
                 show_confirmation(T_('No Problem'), T_('Art has been added'), $burl);
                 break;