+++ lib/class/stream_url.class.php
@@ -65,6 +65,7 @@ class Stream_URL extends memory_object
                     if (make_bool($value)) {
                         $results['type'] = 'video';
                     }
+                // Intentional break fall-through
                 default:
                     // Nothing
                 break;