+++ lib/class/subsonic_api.class.php
@@ -2169,7 +2169,6 @@ class Subsonic_Api
         }
         self::apiOutput($input, $response);
     }
-    /*     * **   CURRENT UNSUPPORTED FUNCTIONS   *** */
 
     /**
      * getChatMessages
@@ -2179,7 +2178,10 @@ class Subsonic_Api
      */
     public static function getchatmessages($input)
     {
-        $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND, '', 'getchatmessages');
+        $since     = (int) $input['since'];
+        $messages  = PrivateMsg::get_chat_msgs($since);
+        $response  = Subsonic_XML_Data::createSuccessResponse('getchatmessages');
+        Subsonic_XML_Data::addMessages($response, $messages);
         self::apiOutput($input, $response);
     }
 
@@ -2189,11 +2191,19 @@ class Subsonic_Api
      * Takes the message in parameter.
      * Not supported.
      */
-    public static function addchatmessages($input)
+    public static function addchatmessage($input)
     {
-        $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND, '', 'addchatmessages');
+        $message = self::check_parameter($input, 'message');
+        $user_id = User::get_from_username($input['u'])->id;
+        if (PrivateMsg::send_chat_msg($message, $user_id)) {
+            $response = Subsonic_XML_Data::createSuccessResponse('addchatmessage');
+        } else {
+            $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND, '', 'addChatMessage');
+        }
         self::apiOutput($input, $response);
     }
+
+    /*     * **   CURRENT UNSUPPORTED FUNCTIONS   *** */
 
     /**
      * getPlayQueue