+++ templates/show_broadcasts_dialog.inc.php
@@ -25,7 +25,7 @@
 <?php
     $broadcasts = Broadcast::get_broadcasts(Core::get_global('user')->id);
     foreach ($broadcasts as $broadcast_id) {
-        $broadcast = new Broadcast($broadcast_id);
+        $broadcast = new Broadcast((int) $broadcast_id);
         $broadcast->format(); ?>
     <li>
         <a href="javascript:void(0);" id="rb_append_dbroadcast_<?php echo $broadcast->id; ?>" onclick="handleBroadcastAction('<?php echo AmpConfig::get('ajax_url') . '?page=player&action=broadcast&broadcast_id=' . $broadcast->id; ?>', 'rb_append_dbroadcast_<?php echo $broadcast->id; ?>');">